portability: fix remaining GCC-blocking pointer/type/label warnings
- wl_menu.c: split ScanNames/ExtScanCodes/ExtScanNames into separate
  declarations; const char* for string tables, byte[] for scan codes
- id_in.h: update IN_GetScanName return type to const char*
- wl_menu.h: fix MenuFadeIn macro to pass gamepal (not &gamepal)
- wl_play.c: fix baseptr = gamepal and VL_SetPalette(gamepal) (not &gamepal)
- wl_menu.c: restore parameter names in CP_* callbacks; add (void) casts
- wl_act2.c: restore ob param in A_Slurpie/T_BJDone; add (void) casts
- wl_inter.c: fix PreloadUpdate signature to word params (matches PM_Preload)
- wl_act2.c/wl_agent.c/wl_draw.c/wl_game.c/wl_inter.c/wl_scale.c/wl_text.c:
  cast signed operands to (unsigned) in mixed signed/unsigned comparisons
- wl_game.c: remove unused restart and startplayloop labels
- wl_menu.c: fix xmin/ymin/xmax/ymax from unsigned to int (match IN_GetJoyAbs)

+++ src/wl_text.c
@@ -166,7 +166,7 @@ void TimedPicCommand(void)
     // wait for time
     //
     TimeCount = 0;
-    while (TimeCount < picdelay)
+    while (TimeCount < (unsigned)picdelay)
         ;
 
     //
@@ -272,7 +272,7 @@ void HandleCommand(void)
         //
         // adjust this line if needed
         //
-        if (px < leftmargin[rowon])
+        if (px < (int)leftmargin[rowon])
             px = leftmargin[rowon];
         break;
     }
@@ -373,7 +373,7 @@ void HandleWord(void)
     //
     VW_MeasurePropString(word, &wwidth, &wheight);
 
-    while (px + wwidth > rightmargin[rowon])
+    while (px + wwidth > (int)rightmargin[rowon])
     {
         NewLine();
         if (layoutdone)
